+++ modules/hanna-utils/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+- feat(ts): Rename `Cleanup<T>` helper, deprecate `Resolve<T>`
 
 ## 0.1.13